fix Contact component name
