+++ package.json
@@ -7,6 +7,7 @@
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
     "google-map-react": "^1.1.5",
+    "node-sass": "^4.13.0",
     "normalize.css": "^8.0.1",
     "page-scroll-to-id": "^1.6.3",
     "react": "^16.12.0",
